Cleanup. Refactor utils in their own dir
