--- api/utils/spotifyDataFilters.js
@@ -1,12 +1,0 @@
-const filterEmptyURL = (article) => { 
-  return article.url && article?.url?.includes('http'); 
-};
-
-const sortTracksByPopularity = (trackA, trackB) => { 
-  return trackB.popularity - trackA.popularity 
-};
-
-module.exports = {
-  filterEmptyURL, 
-  sortTracksByPopularity
-}
